# apps/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.CreateAdminDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/auth.Model'
    type: object
  auth.Model:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      twofa_status:
        type: boolean
      updatedAt:
        type: string
    type: object
  auth.RefreshTokenDto:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  auth.RegisterDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RegistrationStatusDto:
    properties:
      enabled:
        type: boolean
    type: object
  auth.RegistrationStatusResponse:
    properties:
      enabled:
        type: boolean
    type: object
  auth.ResetPasswordDto:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  auth.SetActiveDto:
    properties:
      active:
        type: boolean
    type: object
  auth.TwoFADisableRequestDto:
    description: TwoFADisableRequestDto is used to request 2FA disable
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.TwoFASetupRequestDto:
    description: TwoFASetupRequestDto is used to request 2FA setup
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.TwoFASetupResponseDto:
    description: TwoFASetupResponseDto is used to respond with 2FA setup info
    properties:
      provisioningUri:
        type: string
      secret:
        type: string
    type: object
  auth.TwoFAVerifyRequestDto:
    description: TwoFAVerifyRequestDto is used to verify a TOTP code
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  auth.TwoFAVerifyResponseDto:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  auth.UpdatePasswordDto:
    description: UpdatePasswordDto is used for updating user password
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
...
  utils.APIError:
    properties:
      message:
        type: string
      status:
        type: string
    required:
    - message
    - status
    type: object
  utils.ApiResponse-any:
    properties:
      data: {}
      message:
        type: string
    required:
    - data
    - message
    type: object
...
  utils.ApiResponse-array_auth_Model:
    properties:
      data:
        items:
          $ref: '#/definitions/auth.Model'
        type: array
      message:
        type: string
    required:
    - data
    - message
    type: object
...
  utils.ApiResponse-auth_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/auth.LoginResponse'
      message:
        type: string
    required:
    - data
    - message
    type: object
  utils.ApiResponse-auth_Model:
    properties:
      data:
        $ref: '#/definitions/auth.Model'
      message:
        type: string
    required:
    - data
    - message
    type: object
  utils.ApiResponse-auth_RegistrationStatusResponse:
    properties:
      data:
        $ref: '#/definitions/auth.RegistrationStatusResponse'
      message:
        type: string
    required:
    - data
    - message
    type: object
...
info:
  contact: {}
  title: Peekaping API
paths:
...
  /auth/2fa/disable:
    post:
      consumes:
      - application/json
      parameters:
      - description: 2FA disable request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.TwoFADisableRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Disable 2FA (TOTP) for user
      tags:
      - Auth
  /auth/2fa/setup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 2FA setup request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.TwoFASetupRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TwoFASetupResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Enable 2FA (TOTP) for user
      tags:
      - Auth
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: 2FA verify request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.TwoFAVerifyRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TwoFAVerifyResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.TwoFAVerifyResponseDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Verify 2FA (TOTP) code for user
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Login admin
      tags:
      - Auth
  /auth/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Password update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.UpdatePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Update user password
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Register new admin
      tags:
      - Auth
  /auth/registration-status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_RegistrationStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Get public registration status (no auth required)
      tags:
      - Auth
...
  /settings/registration:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_RegistrationStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Get registration enabled status
      tags:
      - Settings
    put:
      consumes:
      - application/json
      parameters:
      - description: Registration status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegistrationStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Set registration enabled status
      tags:
      - Settings
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-array_auth_Model'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: List all admin users
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: Create admin data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.CreateAdminDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Create a new admin user
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Delete an admin user
      tags:
      - Users
  /users/{id}/active:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Active status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.SetActiveDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-auth_Model'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      - ApiKeyAuth: []
      summary: Set user active status
      tags:
      - Users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - JwtAuth: []
      summary: Reset admin user password
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication (pk_ prefix format)
    in: header
    name: X-API-Key
    type: apiKey
  JwtAuth:
    description: JWT token authentication (Bearer token format)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"